Set the AADJ bit in the EP2ISOINPKTS register. (#37)
* Set the AADJ bit in the EP2ISOINPKTS register.
Without auto-adjustment in high-speed mode the device wants to send
the same amount of data as requested packet size. If there's not enough
data in the FIFO to fill the packet, it fails with EPROTO status marked
in iso packet status.

Enabling auto-adjustment changes that behavior - the device can now
send an empty response or a smaller response than requested (in chunks
of 1024B).

* Store sample rate in a global variable, set AADJ bit if combined sample
rate across channels is higher or equal to 24MHz.

* use efrate to indicate effective rate - ensure that 500kHz is not larger
than 1MHz.

* Store samplerate as a WORD with values in kilosamples/s units.
Store number of channels in a global var (numchannels).